refactor(portal): extract data section into DataSection.svelte
moves the "your data" section (PDS uploads toggle/backfill, media export,
delete-account flow) out of the portal page into a co-located component;
parent shrinks 987 -> 396 lines.
